fix: satisfy backend-size and pkg-file-count CI gates
Split oversized validator, contract, and service test files into existing
package owners, consolidate branch-added interface and projection tests to
restore package file limits, and remove unused agent tool clone helper.
