+++ pkg/service/local_model_runtime_test.go
@@ -189,75 +189,8 @@ func TestInvokeModel_UsesModelHostLeasesAndReusesLoadedRuntime(t *testing.T) {
 }
 
 func TestLoadWorkersFromConfig_LocalModelWorkerUsesManagedRuntimePath(t *testing.T) {
-	audioPath := filepath.Join(t.TempDir(), "speech.wav")
 	provider := &providerCallRecorder{}
-	runtime := &fakeLocalModelRuntime{
-		response: interfaces.InferenceResponse{
-			Content: mustMarshalAudioContentResponse(t, audioPath),
-		},
-	}
-	factoryCfg := localModelFactoryConfig()
-	cache := localModelTestCacheLayout(t)
-	runtimeCfg := newLoadedFactoryConfigForServiceTest(t, "", factoryCfg, localModelRuntimeWorkers(), map[string]*interfaces.FactoryWorkstationConfig{
-		"speak": {
-			Name:           "speak",
-			Type:           interfaces.WorkstationTypeInvoke,
-			WorkerTypeName: "tts-worker",
-			Operation:      "TTS",
-			OperationBindings: []interfaces.ModelOperationBinding{{
-				Slot: "text",
-				Selector: &interfaces.ModelOperationBindingSelector{
-					Label: "utterance",
-					Type:  interfaces.ModelOperationContentTypeText,
-				},
-			}},
-		},
-	})
-
-	opts, err := loadWorkersFromConfig(
-		"",
-		factoryCfg,
-		"",
-		runtimeCfg,
-		nil,
-		logging.NoopLogger{},
-		true,
-		provider,
-		nil,
-		nil,
-		nil,
-		nil,
-		nil,
-		nil,
-		nil,
-		nil,
-		localModelDomain{
-			resources: newLocalModelResourceLimiter(),
-			assets: staticModelAssetPuller{
-				cache: cache,
-			},
-			runtime: runtime,
-			manager: newManagedLocalModelManager(staticModelAssetPuller{
-				cache: cache,
-			}, runtime),
-		},
-	)
-	if err != nil {
-		t.Fatalf("loadWorkersFromConfig: %v", err)
-	}
-
-	fc := &factory.FactoryConfig{}
-	for _, opt := range opts {
-		opt(fc)
-	}
-	exec, ok := fc.WorkerExecutors["tts-worker"]
-	if !ok {
-		t.Fatal("expected tts-worker executor to be registered")
-	}
-	wsExec, ok := exec.(*workers.WorkstationExecutor)
-	if !ok {
-		t.Fatalf("expected *workers.WorkstationExecutor, got %T", exec)
-	}
+	wsExec, runtime := localModelManagedRuntimeWorkerExecutor(t, provider)
 
 	result, err := wsExec.Execute(context.Background(), interfaces.WorkDispatch{
 		DispatchID:      "dispatch-tts",
@@ -483,6 +416,78 @@ func TestNewRecordingModelRunner_LocalModelWorkerEventsStayDetachedFromLaterSour
 	if responsePayload.Resources == nil || len(*responsePayload.Resources) != 1 || (*responsePayload.Resources)[0].Name != "omnivoice-cache" {
 		t.Fatalf("response resources = %#v, want omnivoice-cache", responsePayload.Resources)
 	}
+}
+
+func localModelManagedRuntimeWorkerExecutor(t *testing.T, provider *providerCallRecorder) (*workers.WorkstationExecutor, *fakeLocalModelRuntime) {
+	t.Helper()
+	audioPath := filepath.Join(t.TempDir(), "speech.wav")
+	runtime := &fakeLocalModelRuntime{
+		response: interfaces.InferenceResponse{
+			Content: mustMarshalAudioContentResponse(t, audioPath),
+		},
+	}
+	factoryCfg := localModelFactoryConfig()
+	cache := localModelTestCacheLayout(t)
+	runtimeCfg := newLoadedFactoryConfigForServiceTest(t, "", factoryCfg, localModelRuntimeWorkers(), map[string]*interfaces.FactoryWorkstationConfig{
+		"speak": {
+			Name:           "speak",
+			Type:           interfaces.WorkstationTypeInvoke,
+			WorkerTypeName: "tts-worker",
+			Operation:      "TTS",
+			OperationBindings: []interfaces.ModelOperationBinding{{
+				Slot: "text",
+				Selector: &interfaces.ModelOperationBindingSelector{
+					Label: "utterance",
+					Type:  interfaces.ModelOperationContentTypeText,
+				},
+			}},
+		},
+	})
+	opts, err := loadWorkersFromConfig(
+		"",
+		factoryCfg,
+		"",
+		runtimeCfg,
+		nil,
+		logging.NoopLogger{},
+		true,
+		provider,
+		nil,
+		nil,
+		nil,
+		nil,
+		nil,
+		nil,
+		nil,
+		nil,
+		localModelDomain{
+			resources: newLocalModelResourceLimiter(),
+			assets: staticModelAssetPuller{
+				cache: cache,
+			},
+			runtime: runtime,
+			manager: newManagedLocalModelManager(staticModelAssetPuller{
+				cache: cache,
+			}, runtime),
+		},
+	)
+	if err != nil {
+		t.Fatalf("loadWorkersFromConfig: %v", err)
+	}
+
+	fc := &factory.FactoryConfig{}
+	for _, opt := range opts {
+		opt(fc)
+	}
+	exec, ok := fc.WorkerExecutors["tts-worker"]
+	if !ok {
+		t.Fatal("expected tts-worker executor to be registered")
+	}
+	wsExec, ok := exec.(*workers.WorkstationExecutor)
+	if !ok {
+		t.Fatalf("expected *workers.WorkstationExecutor, got %T", exec)
+	}
+	return wsExec, runtime
 }
 
 func localModelExecutionRecorderFixture(t *testing.T, eventTime time.Time) (*workers.WorkstationExecutor, *factoryevents.FactoryEventHistory, string) {
